refactor: move MovePostprocessor to replication.postprocess
