+++ services/staging/src/main/java/com/composum/sling/platform/staging/replication/postprocess/VersionablePostProcessor.java
@@ -1,4 +1,4 @@
-package com.composum.sling.platform.staging.replication;
+package com.composum.sling.platform.staging.replication.postprocess;
 
 import org.apache.sling.api.resource.Resource;
